Make function names more accurate
The `children` and `parents` labels for graph traversals are not exactly
accurate.

In a tree (graph theory) 'parents' are the nodes that have a directed
edge into the node in question. 'children' are nodes that receive
a directed edge from the node.

The current query service API misuses this vocabulary a bit in that it
calls all nodes in the subtree of a given node 'children' and the nodes
in the path to the root node 'parents'.

This commit fixes this issue.