Format all exceptions in RFC 7807 style (#304)
Previously, any exception not subclassing
connexion.exceptions.ProblemException would be rendered in plain
text. This is not ideal for clients that make the assumption that all
responses are JSON-formatted (including our own query builder UI).

Install a global error handler that formats exceptions as a JSON HTTP
500 response. This roughly matches dss.DSSApp.common_error_handler in
https://github.com/HumanCellAtlas/data-store/blob/master/dss/__init__.py
(but is hopefully simpler).